feat: refactor data model webhook event handling
